Finished implementing cliente and server
The client sends random positions to the server which verifies if the postion corresponds to a dead snake or an obstacle in the board. If the condition is true then the dead snake or obstacle is removed ottherwise nothing happens.
The cliente continues sending random positions until the game ends.